Nimble: bump LastUpgradeCheck to eliminate Xcode proj warnings.
Set the LastUpgradeCheck to `1500` which corresponds to Xcode 15 (that
supposed to be announced in ~1 year from now) to eliminate the
Xcode proj file warnings. These warnings suggest to upgrade the proj
file and raise the deployment target.

This should "buy" at least one year w/o these annoying warnings.

Setting this number to `1500` will allow to face these (and perhaps
other) warnings in next year and decide what we wanna do with them,
since we don't wanna ignore them completely.